src/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.11.13",
    "crawl4ai>=0.7.6,<0.8.0",
    "deepagents>=0.3.12,<0.4",
    "falkordb>=1.2.2",
    "fpdf2>=2.8.7",
    "google-genai>=1.39.1",
    "httpx>=0.28.1",
    "langchain>=1.2.11,<2",
    "langchain-google-genai>=4.2.1,<5",
    "langgraph>=1.1.0,<2",
    "loguru>=0.7.3",
    "minio>=7.2.17",
    "nltk>=3.9.1",
    "numpy>=2.2.6",
    "openai>=1.109.1",
    "openpyxl>=3.1.5",
    "pydantic>=2.11.9",
    "pymilvus>=2.6.5",
    "pyperclip>=1.11.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "pyvi>=0.1.1",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "textual>=6.11.0",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
where = ["src"]
